\name{eu3-package}
\alias{eu3-package}
\alias{eu3}
\docType{package}
\title{
  linking with libacmacsbase test
}
\description{
  linking with libacmacsbase test
}
\details{
  clr = acmacs.Color("brown");
  clr$hex
  clr$name
}
\author{
Eugene Skepner, eu@antigenic-cartography.org.

Maintainer: Eugene Skepner <eu@antigenic-cartography.org>
}
\keyword{ package }
\seealso{
  libacmacsbase
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
  clr = new(acmacs.Color, "brown");
  clr$hex
  clr$name
     ## These can be in \dontrun{} and \donttest{} blocks.
  }
}
